+++ recipe-book-angular/src/app/components/recipe-added-snackbar/recipe-added-snackbar.component.html
@@ -1,3 +1,6 @@
 <div class="flex justify-around items-center">
-    <p>{{"ADDED_RECIPE" | translate}}</p> <a mat-button routerLink='/recipe/{{data}}'><span class="text-zinc-50">{{"TO_RECIPE" | translate}}</span></a>
+  <p>{{ data.text | translate }}</p>
+  <a mat-button routerLink="/recipe/{{ data.recipeId }}"
+    ><span class="text-zinc-50">{{ "TO_RECIPE" | translate }}</span></a
+  >
 </div>
