+++ recipe-book-angular/src/app/components/recipe-added-snackbar/recipe-added-snackbar.component.ts
@@ -7,5 +7,7 @@ import { MAT_SNACK_BAR_DATA } from '@angular/material/snack-bar';
   styleUrls: ['./recipe-added-snackbar.component.css'],
 })
 export class RecipeAddedSnackbarComponent {
-  constructor(@Inject(MAT_SNACK_BAR_DATA) public data: number) {}
+  constructor(
+    @Inject(MAT_SNACK_BAR_DATA) public data: { text: string; recipeId: number }
+  ) {}
 }
