+++ recipe-book-angular/src/app/components/recipe-form/recipe-form.component.css
@@ -1,0 +1,3 @@
+form mat-form-field {
+  width: 100%;
+}
