+++ recipe-book-angular/src/app/pages/add-recipe/add-recipe.component.html
@@ -1,148 +1,8 @@
 <div class="container mx-auto shadow p-3">
   <div class="m-auto">
-    <h1 class="text-3xl font-bold text-center">{{"ADD_RECIPE" | translate}}</h1>
-    <form [formGroup]="recipeForm" (ngSubmit)="onSubmit()">
-      <div>
-        <div formGroupName="recipe">
-          <div>
-            <mat-form-field appearance="fill">
-              <mat-label>{{"NAME" | translate}}</mat-label>
-              <input matInput formControlName="name" name="name" />
-            </mat-form-field>
-          </div>
-          <div fxFlex="100%">
-            <mat-form-field appearance="fill">
-              <mat-label>{{"DESCRIPTION" | translate}}</mat-label>
-              <input matInput formControlName="description" name="description" />
-            </mat-form-field>
-          </div>
-          <div fxFlex="100%">
-            <mat-form-field appearance="fill">
-              <mat-label>{{"COOKING_TIME" | translate}}</mat-label>
-              <input matInput formControlName="cookingMinutes" name="cookingMinutes" />
-              <mat-error>{{"COOKING_TIME_ERROR" | translate}}
-              </mat-error>
-            </mat-form-field>
-          </div>
-          <div fxFlex="100%">
-            <mat-form-field appearance="fill">
-              <mat-label>{{"SKILL_LEVEL" | translate}}</mat-label>
-              <mat-select formControlName="skillLevel" placeholder="Skill level">
-                <mat-option value="EASY">{{"EASY" | translate}}</mat-option>
-                <mat-option value="MEDIUM">{{"MEDIUM" | translate}}</mat-option>
-                <mat-option value="HARD">{{"HARD" | translate}}</mat-option>
-              </mat-select>
-            </mat-form-field>
-          </div>
-        </div>
-        <div formGroupName="category">
-          <div fxFlex="100%">
-            <mat-form-field appearance="fill">
-              <mat-label>{{"CATEGORY" | translate}}</mat-label>
-              <input #inp type="text" placeholder="Pick one" aria-label="Number" matInput formControlName="name"
-                [matAutocomplete]="autoCategory">
-              <mat-autocomplete [panelWidth]="400" #autoCategory="matAutocomplete">
-                <mat-option *ngIf="categorySearchLoading">{{"LOADING" | translate}}</mat-option>
-                <ng-container *ngIf="!categorySearchLoading">
-                  <mat-option *ngFor="let category of filteredCategories" [value]="category">
-                    <p>{{category}}</p>
-                  </mat-option>
-                </ng-container>
-              </mat-autocomplete>
-            </mat-form-field>
-          </div>
-        </div>
-        <div formGroupName="diet">
-          <div>
-            <mat-form-field appearance="fill">
-              <mat-label>{{"DIET" | translate}}</mat-label>
-              <input #inp type="text" placeholder="Pick one" aria-label="Number" matInput formControlName="name"
-                [matAutocomplete]="autoDiet">
-              <mat-autocomplete [panelWidth]="400" #autoDiet="matAutocomplete">
-                <mat-option *ngIf="dietSearchLoading">{{"LOADING" | translate}}</mat-option>
-                <ng-container *ngIf="!dietSearchLoading">
-                  <mat-option *ngFor="let diet of filteredDiets" [value]="diet">
-                    <p>{{diet}}</p>
-                  </mat-option>
-                </ng-container>
-              </mat-autocomplete>
-            </mat-form-field>
-          </div>
-        </div>
-        <div class="shadow p-3 bg-sky-50 rounded-lg flex flex-col mb-3">
-          <h3 class="text-xl text-center">{{"INGREDIENTS" | translate}}</h3>
-          <ng-container formArrayName="ingredientQuantityDtos">
-            <ng-container *ngFor="
-            let ingredientForm of ingredientQuantityDtos.controls;
-            let i = index
-          ">
-              <div class="grid grid-cols-10 gap-4" [formGroupName]="i">
-                <mat-form-field class="col-span-6" appearance="fill">
-                  <mat-label>{{"INGREDIENT_NAME" | translate}}</mat-label>
-                  <input #inp type="text" placeholder="Pick one" aria-label="Ingredient name" matInput
-                    formControlName="ingredientName" [matAutocomplete]="auto">
-                  <mat-autocomplete [panelWidth]="400" #auto="matAutocomplete">
-                    <mat-option *ngIf="ingredientSearchLoading">{{"LOADING" |
-                      translate}}</mat-option>
-                    <ng-container *ngIf="!ingredientSearchLoading">
-                      <mat-option *ngFor="let ingredient of filteredIngredients" [value]="ingredient">
-                        <p>{{ingredient}}</p>
-                      </mat-option>
-                    </ng-container>
-                  </mat-autocomplete>
-                </mat-form-field>
-                <mat-form-field appearance="fill" class="col-span-3">
-                  <mat-label>{{"QUANTITY" | translate}}</mat-label>
-                  <input matInput formControlName="quantity" name="quantity" />
-                </mat-form-field>
-                <button mat-icon-button>
-                  <mat-icon (click)="deleteIngredientQuantityDto(i)">
-                    delete_forever</mat-icon>
-                </button>
-              </div>
-            </ng-container>
-          </ng-container>
-          <button class="m-auto flex-none" mat-mini-fab color="primary" (click)="addIngredientQuantityDto()" type="button">
-            <mat-icon>add</mat-icon>
-          </button>
-
-        </div>
-        <div cdkDropList (cdkDropListDropped)="drop($event)" class="shadow p-3 bg-sky-50 rounded-lg flex flex-col mb-3">
-          <h3 class="text-xl text-center">{{"STEPS" | translate}}</h3>
-          <ng-container formArrayName="stepNumberDtos">
-            <ng-container *ngFor="let stepNumberDto of stepNumberDtos.controls; let i = index">
-              <div class="flex content-center" [formGroupName]="i" cdkDrag>
-                <mat-form-field appearance="fill">
-                  <mat-label>{{"STEP" | translate}}</mat-label>
-                  <input #inp type="text" placeholder="Pick one" aria-label="Number" matInput
-                    formControlName="stepDescription" [matAutocomplete]="auto">
-                  <mat-autocomplete [panelWidth]="400" #auto="matAutocomplete">
-                    <mat-option *ngIf="stepSearchLoading">{{"LOADING" | translate}}</mat-option>
-                    <ng-container *ngIf="!stepSearchLoading">
-                      <mat-option *ngFor="let step of filteredSteps" [value]="step">
-                        <p>{{step}}</p>
-                      </mat-option>
-                    </ng-container>
-                  </mat-autocomplete>
-                </mat-form-field>
-                <button mat-icon-button>
-                  <mat-icon (click)="deleteStepNumberDto(i)">
-                    delete_forever</mat-icon>
-                </button>
-
-              </div>
-            </ng-container>
-          </ng-container>
-          <button class="m-auto" mat-mini-fab color="primary" (click)="addStepNumberDto()" type="button">
-            <mat-icon>add</mat-icon>
-          </button>
-        </div>
-        <div class="flex m-3">
-          <button mat-stroked-button class="m-auto" color="primary" type="submit">
-            {{"SUBMIT" | translate}}<mat-icon>chevron_right</mat-icon>
-          </button>
-        </div>
-      </div>
-    </form>
+    <h1 class="text-3xl font-bold text-center">
+      {{ "ADD_RECIPE" | translate }}
+    </h1>
+    <app-recipe-form [isEdit]="false"></app-recipe-form>
   </div>
 </div>
