+++ recipe-book-angular/src/app/pages/edit-recipe/edit-recipe.component.html
@@ -1,0 +1,12 @@
+<div class="container mx-auto shadow p-3">
+  <div class="m-auto">
+    <h1 class="text-3xl font-bold text-center">
+      {{ "EDIT_RECIPE" | translate }}
+    </h1>
+    <app-recipe-form
+      *ngIf="!isLoading"
+      [isEdit]="true"
+      [recipe]="recipeToAdd"
+    ></app-recipe-form>
+  </div>
+</div>
