+++ recipe-book-angular/src/assets/i18n/en.json
@@ -25,6 +25,8 @@
   "STEPS": "Steps",
   "INGREDIENTS": "Ingredients",
   "ADDED_RECIPE": "Recipe added",
+  "RECIPE_EDITED": "Recipe edited",
   "PREP_TIME": "Prep time",
-  "EDIT": "Edit"
+  "EDIT": "Edit",
+  "EDIT_RECIPE": "Edit recipe"
 }
